+++ django_mobile_app_distribution/settings.py
@@ -29,6 +29,10 @@ MOBILE_APP_DISTRIBUTION_ANDROID_UPLOAD_TO_DIRECTORY_NAME = get(
     'MOBILE_APP_DISTRIBUTION_ANDROID_UPLOAD_TO_DIRECTORY_NAME',
     'android_apps'
 )
+MOBILE_APP_DISTRIBUTION_APP_ICON_DIRECTORY_NAME = get(
+    'MOBILE_APP_DISTRIBUTION_ANDROID_UPLOAD_TO_DIRECTORY_NAME',
+    'app_icons'
+)
 MOBILE_APP_DISTRIBUTION_ANDROID_FILE_STORAGE_PATH = join(
     settings.BASE_PATH,
     get('MOBILE_APP_DISTRIBUTION_ANDROID_FILE_STORAGE_PATH', 'android')
@@ -48,6 +52,9 @@ PLIST_APP_URL = '__app_url__'
 PLIST_BUNDLE_IDENTIFIER = '__bundle_identifier__'
 PLIST_BUNDLE_VERSION = '__bundle_version__'
 PLIST_APP_TITLE = '__app_title__'
+PLIST_DISPLAY_IMAGE = '__display_image'
+PLIST_FULL_SIZE_IMAGE = '__full_size_image'
+
 IOS_PLIST_BLUEPRINT = """<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
@@ -86,3 +93,54 @@ IOS_PLIST_BLUEPRINT = """<?xml version="1.0" encoding="UTF-8"?>
     app_title=PLIST_APP_TITLE
 )
 
+IOS_PLIST_BLUEPRINT_IOS9 = """<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+    <key>items</key>
+    <array>
+        <dict>
+            <key>assets</key>
+            <array>
+                <dict>
+                    <key>kind</key>
+                    <string>software-package</string>
+                    <key>url</key>
+                    <string>{url}</string>
+                </dict>
+            </array>
+            <dict>
+                <key>kind</key>
+                <string>display-image</string>
+                <key>url</key>
+                <string>{display_image}</string>
+            </dict>
+            <dict>
+                <key>kind</key>
+                <string>full-size-image</string>
+                <key>url</key>
+                <string>{full_size_image}</string>
+            </dict>
+            <key>metadata</key>
+            <dict>
+                <key>bundle-identifier</key>
+                <string>{bundle_id}</string>
+                <key>bundle-version</key>
+                <string>{bundle_version}</string>
+                <key>kind</key>
+                <string>software</string>
+                <key>title</key>
+                <string>{app_title}</string>
+            </dict>
+        </dict>
+    </array>
+</dict>
+</plist>
+""".format(
+    url=PLIST_APP_URL,
+    bundle_id=PLIST_BUNDLE_IDENTIFIER,
+    bundle_version=PLIST_BUNDLE_VERSION,
+    app_title=PLIST_APP_TITLE,
+    display_image=PLIST_DISPLAY_IMAGE,
+    full_size_image=PLIST_FULL_SIZE_IMAGE
+)